+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 
 Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image `hardisgroupcom/sfdx-hardis@beta`
 
+- Handle bug when a branch .sfdx-hardis.yml config file is empty
+- Upgrade default API version to 61
+- Additional log when generating manifest package.xml from org
+- Add error tip: Network issue (ECONNABORTED, ECONNRESET)
+
 ## [4.50.1] 2024-07-29
 
 - Fix report file name of [hardis:org:monitor:limits](https://sfdx-hardis.cloudity.com/hardis/org/monitor/limits/)